fixed bugs in acmg_sf report script
