+++ workflow/rules/acmg_sf.smk
@@ -7,7 +7,7 @@ acmg_sf_input_report_type = [
 if len(children) > 0:
     acmg_sf_input_report_type.append("wgs.denovo.CH")
 
-rule add_acmg_sf_column:
+rule add_acmg_sf_columns:
     input:
         report="reports/{family}.{input_report_type}.hg38.csv",
         acmg_sf_list=config["annotation"]["general"]["acmg_sf_list"],
@@ -23,4 +23,22 @@ rule add_acmg_sf_column:
     conda:
         "../envs/acmg_sf.yaml"
     script:
-        "../scripts/add_acmg_sf_annotation.py"
+        "../scripts/add_acmg_sf_columns.py"
+
+rule create_acmg_sf_report:
+    input:
+        reports=lambda wildcards: expand(
+            "reports/{family}.{input_report_type}.SF.hg38.csv",
+            family=wildcards.family,
+            input_report_type=acmg_sf_input_report_type,
+        ),
+    output:
+        report="reports/{family}.acmg_sf_report.hg38.csv",
+    params:
+        acmg_sf_version=config["annotation"]["general"]["acmg_sf_version"],
+    log:
+        "logs/report/acmg_sf/{family}.acmg_sf_report.log",
+    conda:
+        "../envs/acmg_sf.yaml"
+    script:
+        "../scripts/create_acmg_sf_report.py"
